+++ suites/selftad/ipstack/ip4_recv_send.c
@@ -69,7 +69,7 @@ main(int argc, char *argv[])
     char ta[32];
     char *agt_a = ta;
     char *agt_b;
-    unsigned  len = sizeof(ta);
+    size_t  len = sizeof(ta);
 
     te_bool enum_iterator;
